+++ src/model.rs
@@ -4,5 +4,5 @@ use serde::{Deserialize, Serialize};
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct UserModel {
   pub _id: String,
-  pub coins: i32
+  pub coins: i64
 }
